Consolidate commands relating with kyverno plugin
